backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
slowapi>=0.1.9
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-google-genai>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
PyPDF2>=3.0.1
httpx>=0.27.0
python-multipart>=0.0.12
